requirements-agent.txt
PyAudio==0.2.14
pygame==2.6.1
SpeechRecognition==3.12.0
python-dotenv==1.0.1
gTTS==2.5.4
langchain==0.2.0
langchain-core==0.2.0
langchain-google-genai==0.0.11
scipy==1.15.2
numpy==1.26.4